validatePSA and parsePSA working
